+++ src/entropy_based/generalized_dim.jl
@@ -46,12 +46,12 @@ The source code of this function is only 3 lines of code.
 function generalized_dim(X::AbstractStateSpaceSet, sizes = estimate_boxsizes(X);
         base = 2, q = 1.0, show_progress = envprog(),
     )
-    H = _threaded_entropies(X, sizes, q, show_progress)
+    H = _threaded_entropies(X, sizes, q, base, show_progress)
     x = -log.(base, sizes)
     return linear_region(x, H)[2]
 end
 
-function _threaded_entropies(X, sizes, q, show_progress)
+function _threaded_entropies(X, sizes, q, base, show_progress)
     progress = ProgressMeter.Progress(length(sizes);
         desc = "Entropy calculation: ", enabled = show_progress
     )